refactor: move historyController method to gameSessionController delete historyController
