--- src/main/java/com/scriptopia/demo/controller/HistoryController.java
@@ -1,33 +1,0 @@
-package com.scriptopia.demo.controller;
-
-import com.scriptopia.demo.service.HistoryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/users/my-page")
-@RequiredArgsConstructor
-public class HistoryController {
-    private final HistoryService historyService;
-
-    /**
-     * 현재는 userId, sessionId를 통해 저장하는데
-     * 인증 관리 부분 끝나면 header에 token 꺼내오고 requestparameter session_id로 저장하게 수정
-     */
-    @PostMapping("/{sid}/history")
-    public ResponseEntity<?> addHistory(@PathVariable String sid, Authentication authentication) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        return historyService.createHistory(userId, sid);
-    }
-
-    /** 개발용: 로컬 MongoDB에 더미 세션 한 건 심어서 테스트용 ObjectId 반환 */
-    @PostMapping("/history/seed")
-    public ResponseEntity<?> seed(Authentication authentication) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        return historyService.seedDummySession(userId);
-    }
-}
